Convert wasi-http proxy example to Rust
Essentially C/C++ support in WASI tools is actually C support and not
C++ in a meaningful sense (e.g., the only thing from C++ there is that
it could be built with C++ compiler and linked with C++ code, but all
the generated code basically plain C).

While I don't mind C APIs in principle, bindings for more or less
complicated WIT descriptions are just not usable without a large library
of helper functions to manage all the boilerplate.

I don't feel like creating such a library for exploratory work I'm
doing. So to avoid getting distracted from the actual work I want to do,
I'm switching to Rust.

It's a shame really, I don't understand why they decided to provide such
a bad C++ interface in the bindings that basically renders it unusable.
Looking at this, I feel like maybe a better way to go about it is not to
generate bad bindings, but to describe properly what is the actual ABI
and let folks create bindings that are actually in line with the
pragamatics of the language.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>