Add skeleton bazel configuration to build WebAssembly
While Cargo component tools work, I think I'm looking for
a solution that plays well with different languages (e.g., Rust
and C++ as a minimum, because that's the languages for which
proxy-wasm has officially supported SDKs).

It seems like using Bazel, which supports both Rust and C++ is
more convenient rather than using Cargo component tools.

I also appreciate the fact that using Bazel gives me a bit more
exposure to the low-level details of how things are built and
linked which is useful learning for my experiment.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>