Tell Bazel to stop building intermediate static libraries
By default it looks like Bazel builds static libraries for each of
the cc_library target. So when Bazel builds cc_library it needs to
do two things:

1. generate object files from the sources
2. generate static library from the object files.

The generated static library is later used to build the final executable.

However, we can tell Bazel to not build static libraries and instead
link all the object files directly without creating the object files.
And that is allegedly faster, so let's do that.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>