Add support for WIT packages to Bazel
We need WIT packages to later generate bindings with wit-bindgen.
Without Bazel I would use wit-deps to handle all the dependencies,
but Bazel, as a build system, is designed to keep track and manage
dependencies, so it seems better to just leave that job to Bazel
when using Bazel.

The support for WIT packages basically boils down to teaching Bazel
how to generate a directory that wit-bindgen expects to see when it
generates bindings (e.g., top level wit files at the root of the
dir and all the required dependencies are in deps/).

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>