Add a simple do-nothing proxy implementation on top of wasi-http
First of all, I didn't actually test that this proxy works yet (e.g.,
I didn't load it into wasmtime and verified that it can proxy some
requests).

With that out of the way, this proxy is very simplistic in multiple
ways:

1. It does not actually implement the state machine and buffering that
   proxy-wasm, not to mention Envoy do - I will have to implement that
   in the future
2. It basically panic on any error it sees - I will have to fix this in
   the future
3. It does not actually modify anything when proxying request/responses
   and does not call proxy-wasm - I will have to fix this in the future
4. It does not actually have any configuration capabilities that normal
   proxies, like Envoy, have, so for now it can only send the requests
   to the destination specified in the request header.

I don't really think I'm going to fix the point 4 in the future, because
ultimately the destination of the http request/response after it was
handled by a filter is something that Envoy (or another proxy) should
figure out.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>