Make proxy example work with wasmtime serve
There were a couple of issues that affected the example:

1. When returning a string property we have to allocate it and
   cannot use statically allocated string; that's because proxy-wasm
   will later try to free the string which will cause problems if it
   wasn't allocated
2. blocking_write_and_flush doesn't actually do flush, it at least
   when sending output to terminal, performs so-called line buffering,
   where it waits for the '\n' in the input to flush.

One question that I don't think I know the answer to is where there
is an essential difference between calling proxy-wasm malloc or using
Rust built-in methods for allocating memory. In other words, do they
use the same underlying memory allocator (if they don't that would
cause a ton of issues) and whether there are any abstractions in the
way between C++ new/Rust allocator and the low-level underlying allocator.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>